redisedge/requirements.txt
# git+https://github.com/beasteers/redis-py.git@fix-closed-connection-writer#egg=redis
redis>=4.2.0
numpy
opencv-python>=4.0.1.23
Pillow
fire
onnxruntime